fullAutoMode now works as expected in landscape on iPhone Simulator
Before this change, fullAutoMode always forced portrait mode. This doesn't make sense when the iphone is in landscape and makes the app do a funny dance where it sometimes goes to portrait and back to landscape without the device changing orientation. This change resolves this by maintaining the real device orientation and restoring it when fullAutoMode is pressed.